# swagger/swagger.yaml
swagger: '2.0'
info:
  description: Aplicación dirigida a compañías de Marketing para administrar todos los contactos de sus clientes para sus campañas.
  version: 1.0.0
  title: Data Warehouse
  contact:
    email: solmartinezpisani@gmail.com
...
schemes:
  - https
  - http
paths:
  /usuarios/registro:
    post:
      tags:
        - usuarios
      summary: Crear usuario
      operationId: createUser
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: |-
            Crear usuario. Perfil puede ser: 
             1 (administrador),  
             0 (perfil básico)
          required: true
          schema:
            $ref: '#/definitions/Register'
      responses:
        '201':
          description: Successful Register
          schema:
            $ref: '#/definitions/RegisterResponse'
        '400':
          description: Unsuccessful Register
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 400
              message:
                type: string
                example: User already exists
  /usuarios/login:
    post:
      tags:
        - usuarios
      summary: Iniciar sesión
      operationId: loginUser
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Login usuario.
          required: true
          schema:
            $ref: '#/definitions/Login'
      responses:
        '200':
          description: Successful login
          schema:
            $ref: '#/definitions/LoginResponse'
        '400':
          description: Unsuccessful login
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 400
              message:
                type: string
                example: Incorrect user or password
  /usuarios:
    get:
      tags:
        - usuarios
      summary: Ver todos los usuarios (admin)
      operationId: getUsers
      produces:
        - application/json
      parameters:
        - name: User token
          in: header
          required: true
          type: string
          description: User JWT token (Bearer)
      responses:
        '200':
          description: Successful operation
          schema:
            type: array
            items: 
              $ref: '#/definitions/GetUsers'
        '400':
          description: Unsuccessful
          schema:
            $ref: '#/definitions/Authenticate'
        '403':
          description: Forbidden
          schema:
            properties:
              status_code:
                type: integer
                format: int64
                example: 403
              message:
                type: string
                example: Forbidden
        '404':
          description: Not Found
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 404
  /usuarios/{id}:
    get:
      tags:
        - usuarios
      summary: Ver un sólo usuario (admin)
      operationId: getUser
      produces:
        - application/json
      parameters:
        - name: User token
          in: header
          required: true
          type: string
          description: User JWT token (Bearer)
        - in: path
          name: id
          type: integer
          required: true
          description: Numeric ID of the user.
      responses:
        '200':
          description: Successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/GetUser'
        '400':
          description: Unsuccessful
          schema:
            $ref: '#/definitions/Authenticate'
        '403':
          description: Forbidden
          schema:
            properties:
              status_code:
                type: integer
                format: int64
                example: 403
              message:
                type: string
                example: Forbidden
        '404':
          description: Not Found
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 404
    patch:
      tags:
        - usuarios
      summary: Editar usuario (admin)
      operationId: updateUser
      produces:
        - application/json
      parameters:
        - name: User token
          in: header
          required: true
          type: string
          description: User JWT token (Bearer)
        - in: path
          name: id
          type: integer
          required: true
          description: Numeric ID of the user to update.
        - in: body
          name: body
          description: >-
            Editar usuario
          required: true
          schema:
            $ref: '#/definitions/UpdateUser'
      responses:
        '200':
          description: Successful operation
          schema:
            properties:
              message:
                type: string
                example: Usuario editado
        '400':
          description: Unsuccessful
          schema:
            $ref: '#/definitions/Authenticate'
        '403':
          description: Forbidden
          schema:
            properties:
              status_code:
                type: integer
                format: int64
                example: 403
              message:
                type: string
                example: Forbidden
        '404':
          description: Not Found
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 404
    delete: 
      tags:
        - usuarios
      summary: Eliminar usuario (admin)
      operationId: deleteProduct
      produces:
        - application/json
      parameters:
        - name: User token
          in: header
          required: true
          type: string
          description: User JWT token (Bearer)
        - in: path
          name: id
          type: integer
          required: true
          description: Numeric ID of the user to delete.
      responses:
        '200':
          description: Successful operation
          schema:
           properties:
              message:
                type: string
                example: Usuario eliminado
        '400':
          description: Unsuccessful
          schema:
            $ref: '#/definitions/Authenticate'
        '403':
          description: Forbidden
          schema:
            properties:
              status_code:
                type: integer
                format: int64
                example: 403
              message:
                type: string
                example: Forbidden
        '404':
          description: Not Found
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 404
...
definitions:
  Authenticate:
    type: object
    properties:
      message:
        type: string
        example: Header must contain JWT to verify identity.
  Register:
    type: object
    properties:
      nombre:
        type: string
        example: Sol
      apellido:
        type: string
        example: Martinez
      email:
        type: string
        example: solmartinezpisani@email.com
      contrasenia:
        type: string
        example: password!13
      admin:
        type: integer
        format: int64
        example: 1
  RegisterResponse:
    type: object
    properties:
      message:
        type: string
        example: Nuevo usuario creado
  Login:
    type: object
    properties:
      email:
        type: string
        example: solmarti1@gmail.com
      contrasenia:
        type: string
        example: password!13
  LoginResponse:
    type: object
    properties:
      token:
        type: string
        example: >-
          eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyTmFtZSI6InVzdWFyaW9kZW1vMjMiLCJpc0FkbWluIjowLCJpYXQiOjE1ODg1NTA2MjgsImV4cCI6MTU4ODYzNzAyOH0.OXF880nyOkmH31U5ZzZaMlfN9EaXCGrrPgECvq-mktk
      admin: 
        type: integer
        format: int64
        example: 1
  GetUsers:
    type: object
    properties:
      id:
        type: integer
        format: int64
        example: 1
      nombre:
        type: string
        example: Sol Martinez
      email:
        type: string
        example: solmartinezpisani@email.com
      admin:
        type: integer
        format: int64
        example: 1
  GetUser:
    type: object
    properties:
      id:
        type: integer
        format: int64
        example: 1
      nombre:
        type: string
        example: Sol Martinez
      email:
        type: string
        example: solmartinezpisani@email.com
      contrasenia: 
        type: string
        example: 4d186321c1a7f0f354b297e8914ab240
      admin:
        type: integer
        format: int64
        example: 1
...
  UpdateUser:
    type: object
    properties:
      nombre:
        type: string
        example: Sol Martinez
      email:
        type: string
        example: solmartinezpisani@email.com
      contrasenia: 
        type: string
        example: 4d186321c1a7f0f354b297e8914ab240
      admin:
        type: integer
        format: int64
        example: 1
...